Pin GQL Dependency
GQL the GraphQL Implementation used within seer-py, was broken by an upgrade to the graphql-core package. This broken the interface of gql that seerpy depends on. Unfortunately `gql` is no longer mainted actively and the most active maintainer doesn't have permission to publish to PyPI, thus why this depends on the `gql` github project itself, which was fixed in https://github.com/graphql-python/gql/pull/52.

https://github.com/graphql-python/gql/issues/51