Fix data reshaping error for irregular EDF uploads
Basically a copy/paste of a similar fix in seer-app. With the
bug in seer-worker fixed, it's technically possible to fix this
by re-uploading all the affected EDF files, but ain't nobody got
time for that.

I should probably refactor the function and test the bug fix,
but I just don't have the willpower for it.